move unique check into component
